backend/requirements.txt
# Core deep learning libraries
torch
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cpu

# Data manipulation and machine learning
pandas
scikit-learn
Pillow

# Web API framework
fastapi
uvicorn
python-multipart
fastapi-mail
python-dotenv
mysql-connector-python
